[Chore] `fakeAsync.test.ts`
